Improve server config and update README (#334)
Onboarding several people reminded me how hard it is to get the Flask and FastAPI servers running locally. 

Fixes #335 

This commit:

* Makes dependency on calling Google Secrets in local mode optional
* Updates README to provide comprehensive instructions to install and run locally
* [Tidy up]: Moves some scripts away from the root directory into sub-directories, either into `scripts/` or `fast-api`
* Updates the app.yaml files to reflect the new location of the FastAPI server script in `fast-api/main.py`
* Adds a new module with a `Config()` class that sets config variables consistently and also sets up logging (see below)

### The Config class

This class looks for needed global variables in three locations in order:

1. Environment variables
2. A `.env` file
3. Google Secrets

If the variable is not found in each location it tries the next. The class also includes a method to validate that the correct variables are found for each server type, and a method to set-up logging or Google Cloud logging, depending on the environment.

This module removes a big chunk of conditional statements at the beginning of both server files, and makes config more consistent.
* Make get_secret() call optional in main.py

* Updated README

* Moving files and fixing relative imports

* Config system

* Revert some relative imports

* README edit

* README edit

* README edit

* README edits

* README edits

* README edit

---------

Co-authored-by: matweldon <dr.mat.weldon@gmail.com>