500 HTTP Response for empty json response from FastAPI (#342)
### What?

Sometimes the backend Fastapi server sends a 200 response but the ranking list associated with each of the inputs is empty. This change adds additional Pydantic checks to the FastAPI response body to ensure that a 503 response is sent when this happens, instead of a 200 response.

### Why?

Currently the 200 response causes Flask to send empty response data to the front end, when actually the request should be retried.

### How?

Adding a validator decorator to the Pydantic Response object that models and validates the ranking/response json object allows us to explicitly check for the empty list problem and return the appropriate 500 response and nice details.
* added custom pydantic check for empty response bodies

* updated NaN to nan in test files as per new numpy versions

* updated even more nan syntax